fix issues in esp32s2 and esp32s3 USB0 endpoint control registers
Update the endpoint control registers for the 6 IN and OUT endpoints:
endpoint0 is slightly different from the others, but the remaining 5
endpoints are all the same.  This updates the DIEPCTL and DOEPCTL
registers for endpoints 2 through 6 to use the same type as endpoint 1
rather than being completely separate types.

This also renames the fields to drop the endpoint number from the name,
and remove the `D_`, `DI_`, and `DO_` prefixes from the fields.  Note
that the base SVD uses these prefixes inconsistently, and not all
endpoints used the same prefixes for all fields.

Additionally, this fixes a few bugs in the field definitions:

- The MPS bit width was incorrectly set to 2 in DIEPCTL.
  This field is 2 bits wide for endpoint 0, but I believe it should be
  11 bits for all other endpoints.

- The EPDIS field was marked read-only in DOEPCTL.  I believe this field
  should be read-only for endpoint 0 (since this endpoint cannot be
  disabled), but should be read-write for other endpoints.

In the future it might be nice to update the DIEPCTL0 and DOEPCTL0
registers to use the same field naming scheme, so that it is consistent.
(e.g., drop the `0` suffix from all fields, and the `D_`, `DI_`, `DO_`
prefixes where those are used.)  However, I didn't include that change
as part of this commit, just to avoid making too many changes at once.
I believe the DOEPCTL0.EPENA0 field is also potentially supposed to be
write-only rather than read-write, but this is a minor difference.

In the future it would potentially also be nice to also group the
endpoint 2-5 registers together in an array to make them easier to
access, as is done in the ESP-IDF C headers.  (It is a little awkward
that endpoint 0 doesn't quite have the same field definitions, meaning
we would need to have endpoint registers followed by a separate array
for the remaining endpoints.)