usb: cluster endpoint registers
This groups the USB endpoint registers into clusters.

Note that endpoint 0 is different from the other endpoints, so it
unfortunately must be put into a separate cluster of its own, for both
the in and out endpoints.  This does produce some ergonomic issues:
The `in_ep0()`, `in_ep1()`, ... `in_ep6()` functions behave as you would
expect.  However, there is also an `in_ep(n: usize)` function, but it's
behavior is slightly surprising, as it only returns data for endpoints
1-6, and the index it accepts is therefore off by one (`in_ep(1)`
returns the same result as `in_ep2()`).