usb: unify endpoint register types where possible
Endpoint 0 has different register definitions than the other endpoints
for some but not all of its registers.  For the registers that are the
same between all endpoints, use `deriveFrom` to make endpoints 1-6 share
the same register type as endpoint 0.  This makes it easier to write
code using these registers that doesn't care if it is operating on
endpoint 0 or not.