fix issues in esp32s2 and esp32s3 USB0 endpoint control registers (#213)
* usb: group DIEPTXF registers into an array

* usb: cluster endpoint registers

This groups the USB endpoint registers into clusters.

Note that endpoint 0 is different from the other endpoints, so it
unfortunately must be put into a separate cluster of its own, for both
the in and out endpoints.  This does produce some ergonomic issues:
The `in_ep0()`, `in_ep1()`, ... `in_ep6()` functions behave as you would
expect.  However, there is also an `in_ep(n: usize)` function, but it's
behavior is slightly surprising, as it only returns data for endpoints
1-6, and the index it accepts is therefore off by one (`in_ep(1)`
returns the same result as `in_ep2()`).

* usb: fix bugs in the DIEPCTL, DOEPCTL, DIEPTSIZ, and DOEPTSIZ registers

These registers were defined incorrectly for endpoints 1 through 6.
This fixes the bid-widths and read-write access modes for several fields
which had their values copied from the endpoint 0 register definitions.

* usb: rename the endpoint 0 control register fields

Rename the register fields for endpoint 0 to match the fields in the
register cluster created for endpoints 1 through 6.

* usb: clean up remaining endpoint register field names

Update the register field names to be consistent between endpoint 0 and
the other endpoints.  This removes the endpoint number suffix, and drops
the `D_`, `DI_`, and `DO_` prefixes that were inconsistently applied to
the field names.

* usb: add registers to access the FIFOs

* usb: specify that the FIFOs allow writing any possible value

This allows users to write any value to the FIFOs without needing to use
the unsafe `bits()` method.

* Upgrade to svdtools 0.3.14

This release contains fixes for handling write constraints as
well as derived registers, which are needed for some of my USB0
peripheral fixes.

* usb: add write constraints to register fields

Add write constraints to individual register fields. This allows
svd2rust to generate safe accessors for these fields, rather than only
being able to access them through unsafe methods.

* usb: unify endpoint register types where possible

Endpoint 0 has different register definitions than the other endpoints
for some but not all of its registers.  For the registers that are the
same between all endpoints, use `deriveFrom` to make endpoints 1-6 share
the same register type as endpoint 0.  This makes it easier to write
code using these registers that doesn't care if it is operating on
endpoint 0 or not.

* usb: add write constraints for fields in the dcfg register

This allows svd2rust to make writes to these fields safe.